+++ src/main.cpp
@@ -26,14 +26,15 @@
  * SOFTWARE.
  */
 
-#include <QApplication>
+#include "Core/Application.hpp"
 #include "GUI/MainWindow.hpp"
 
+using namespace ShaderIDE;
 using namespace ShaderIDE::GUI;
 
 int main(int argc, char** argv)
 {
-    QApplication app(argc, argv);
+    Application app(argc, argv);
     MainWindow window;
     window.show();
     return app.exec();